added compensation and claw functions
I added a simple PID based compensation to driver control. I added a few
compensation based commands to the auto library. I added a kill switch
to the compensation, and you should be careful to always use it before
manually controlling the claw in auto. From now on all claw functions in
auto need to call it at the beginning. I also added a function to move
the claw to a given position and then compensate for auto. Also praise
lord Evan for the number -1.